Fix for DM failure: Std subcloud installation
Playbook fails when it tries to reboot and subsequent reboot was not executed.
Required conditions were added to execute the second reboot task.

Test plan:
Deployment of subcloud3 was failing at the reboot task. With this fix,
verified the reboot task is getting failed and subsequent reboot is executed
successfully.

Signed-off-by: mmanthir <mahalakshmi.manthiram@windriver.com>