Fixing storage.filesystem syncrhonization
The Storage.Filesystem synchronization is failing for AIO-DX because
the 'ceph' host filesystem is created during the first unlock of the
controllers. The host profile must not have the 'ceph' filesystem, but
the current config will have it after the unlock, because the sysinv
will create that automatically to support the 3 Ceph monitors.

This fix will ignore the 'ceph' filesystem synchronization when this
is an AIO-DX deployment.

Test-Plan
  PASS: Deploy an AIO-DX and check both controllers are InSync and
        Reconciled
  PASS: Deploy a Standard and check all the hosts are InSync and
        Reconciled

Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>