Fix host availability update post reconciled
In cases where the host is already insync and reconciled, if it was
degraded before the unlock, the availability status was not being
updated.

This commit adds a check in the Reconcile method to ensure that the
availability status it's updated if the host is already reconciled but
not available.

Test Plan:
PASS: Build image and deploy a AIO-DX where the issue was seen. Verify
that the 'degraded' status it's updated to 'available' after unlock.

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
(cherry picked from commit 031b3b047368910ad97025fa9de139bdedfa57b6)