Include all status of hosts except online
Currently, the DM only checks for the offline state, not the
power-off state, when hosts are not in simplex mode.
This change ensures that all host statuses, except online, are included.

Test-Plan
PASS: Deploy AIO-DX on VDM
PASS: DM adding the controller-1, check the controller-1 is added into the "system host-list",
power off the controller-1 using "system host-power-off controller-1
Signed-off-by: sbhardwa <sharang.bhardwaj@windriver.com>