Support static IP assigment on all interfaces
This commmit removes the restriction of static IP assigment on
interface types, allows static IP being assigned on not only the data
interfaces. This change aligns with the restriction of the sysinv API.

Test Plan:
Passed: freshly deploy an AIOSX system.
Passed: verified a static IP can be assigned on platform interface.

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>