Refresh networks in host info post reconciliation
After merging the platform network reconciliation into the host
controller, platform networks creation are delayed if not
pre-configured. This commit forces to update the networks in the
host info after reconciling the platform networks, prevents interfaces
network association missed in the following steps.

Test plan:
Create a storage-net as platform network in the DM configuration,
verfied the interface association to this network done before
unlocking the controller.

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>