Adapt customed ns and improve error logging
This commit checks namespace from the configuration file to prevent
user using customzied namespace that breaks the playbook.

In addition, this commit updates the error checks on multinode systems
that only failed when host is not unlocked and not in online or
available status, which indicates the host was not ready for the
deployment manager to configure it.

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>