Cherry picking fixes to 24.09.4 (#447)
* Refresh networks in host info post reconciliation (#441)

* Extend webhook timeout to 30s in helmcharts

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>

* Refresh networks in host info post reconciliation

After merging the platform network reconciliation into the host
controller, platform networks creation are delayed if not
pre-configured. This commit forces to update the networks in the
host info after reconciling the platform networks, prevents interfaces
network association missed in the following steps.

Test plan:
Create a storage-net as platform network in the DM configuration,
verfied the interface association to this network done before
unlocking the controller.

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>

---------

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>

* Bug-fix: avoid IPv4 mode overridden by IPv6 mode (#442)

The IPv4 mode and IPv6 mode are assigned with the mode address,
updating the mode's value can cause IPv4 mode being overridden by IPv6
mode. This commit fixed this issue by assigning them with different
variables' address.

Test plan:
Verified interface can be assigned with IPv4 address, the interface's
IPv4 mode is updated to static.

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>

* Fix ready state notification race condition in system controller (#443)

When system controller is ready, it will notify all the other reconcilers
about the ready state, so the other controllers can start the
reconciliation loop. This commit ensures that the system controller
can notify the other controller even if the last iteration failed to
call the webhook, but the system is reconciled.

Test Case:
- PASS: Deploy AIO-SX

* Enable the creation of two data VLAN with same VID (#445)

This commit fixes the vlan interface search, searching the interface by
name and VID to allow the creation of two VLAN with same VID when
using two different physical interfaces.

Test Case:
- PASS: configure and verify two VLAN using the same VID

---------

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Co-authored-by: Yuxing Jiang <yuxing.jiang@windriver.com>