Fix get administrative state tasks (#472)
This commit fixes the logic to retrieve the administrative state,
search on host resource and host profile resource. And the playbook
fails if the unable to retrieve the desired final administrative
state.

In a SX subcloud, if the the desired final configured state on
host/hostprofile is "unlocked", the playbook will wait the host to be
in an unlocked administrative state, otherwise it will skip the
verification.

Test Plan:
- PASS: playbook executation must wait the host be unlocked when:
  1) administrativeState: "unlocked" is configured only on the host
     resource.
  2) administrativeState: "unlocked" is configured only on the
     host profile resource.
  3) administrativeState: "unlocked" is configured on the host
     resource and administrativeState: "locked" on the host profile
     resource.
- PASS: playbook execution must skip unlock verification when:
  1) administrativeState: "locked" is configured on host profile
     resource.
  2) administrativeState: "locked" is configured on host resource.

Failure Path:
- PASS: administrativeState not configured on host resource or host
        profile resource.
- PASS: playbook execution unable to retrieve administrativeState.

Signed-off-by: Wallysson Silva <wallysson.silva@windriver.com>