Delete stale VF and VLAN if name changes
This commit updates the methods that avoids find the VLAN and VF if
its name changes. Enables deleting a VF or a VLAN if its name changes.
The main reason is the system doesn't allow VLAN or VF with duplicated
names, if an existing VF or VLAN is expected to change its name and
the name is expected to be assigned to another VLAN or VF, the staled
one needs to be deleted.

This commit doesn't include the name check on the AE interfaces,
remains the existing checks unless further support expected and
the logic clarified.

Tests:
Passed - verified a VLAN name can be assigned to a different VID.
Passed - verified a VF name can be assgined to a different lower
interface.

Signed-off-by: Jiang, Yuxing <Yuxing.Jiang@windriver.com>